Move MiqAeMethodService::MiqAeServiceModelLegacy to the correct path
